### # GSD-21, GSD-24 : US05.02 Documentation recommendations ###
General suggestions:

	1. Move the Detailed Examples section up from Section 8 to become
Section 4---this is how you suggest readers use the document, so the structure
should follow that advice. Since this is intended as a user's manual, these
actual usage suggestions should come first.

	2. Replace the Syntax, Object Types, and Formatted Output sections
with the actual help files. These are essentially reference information, and
thus would tend to duplicate the contents of the help files. (You should ensure
that no information is lost, however.)
 
	3. It would be good to have a little bit of intro material to say 
just what a spending function is---the section Spending Functions assumes a 
certain degree of familiarity with the concept and the formal description in 
Statistical Methods comes too late to help orient a reader who may be familiar 
with the Wang-Tsiatis family of designs but with no previous experience with 
spending functions. (Maybe just a paragraph or two, similar to the intro to 
the spending function chapter in Jennison and Turnbull.) [You may feel this is 
unnecessary---since all I know of sequential trials is a long-ago reading of 
the S+SeqTrial manual and an hour or two in Jennison and
Turnbull, I have no subject matter expertise. But _brief_ recaps that help 
orient the reader to what's going on are usually welcome in
tutorial documents.]



	4. In the .tex file gsDesignManualPortable.tex, you will find a few 
comments starting %RBC--these are usually requests for information. 
On my hardcopy of the original PDF, I noted that I preferred the formatting of 
the gsDesign() syntax section to that of all subsequent syntax sections, and 
much of my line editing was intended to bring other sections into line 
with that. What I liked about that first section was that for each argument, 
the first thing shown was allowable input. In some of the later sections, 
I couldn't always tell, so many of the %RBC comments are requests for that 
info. Some of the others point out things that looked to errors like me, or 
missing info, or whatnot, but I couldn't be sure. So please take a look and 
see if any action needs to be taken on those.

	5. Some stylistic suggestions:

	a. Use code font \texttt{} consistently for all code-related items
(functions, classes, objects, parameters, arguments)
	b. Use math mode consistently for equations and inequalities, with
\mathtt as necessary to comply with item a.
	c. Use LaTeX style cross-references rather than hard-coded ALLCAPS.
	d. When using a "dash" as punctuation, ensure you are using em-dashes
(---) and not en-dashes (--). En-dashes are appropriate for specifying ranges
(1--10).
	e. As much as possible, write in the present tense.
	f. avoid the abbreviations i.e. and e.g., because too many people
don't know what they mean. Just spell out "that is" and "for example".


* Pg. 6, delta arg description: Use '\texttt{delta} >= 0' rather than '\delta > 0' to be consistent. 
Notice the inclusion of 0 as later in that paragraph it says that delta=0 is acceptable.

* Pg. 16, Table 1, Spending function definitions and parameterizations.
For sfHSD(), range of gamma is given as [-40, 40).
However, code error messages specify "must be <=40 in absolute value", implying [-40, 40].

* Pg. 7: upper limit not given for k; however, code checks that it is <=30.

* Pg. 1: add serial comma after "fully characterize"

* Pg. 2: modify list to have parallel structure by beginning each item with "The xxx function returns"

* Pg. 3, item 3 in gsDesign list: add a hyphen between "three" and "parameter"

* Pg. 5 (and elsewhere): Use {\bf Note:} instead of "NOTE:"

* Pg. 5 (and elsewhere): "2-sided" should be spelled out "two-sided".
Likewise "2-parameter" as "two-parameter", and so on.

* Pg. 5 (and elsewhere): "1st" should be spelled out "first"

* Pg. 6 (and elsewhere): use math mode to obtain real minus signs; don't use 
hyphens as minus signs. (For example, $1 - \mathtt{alpha}$, not 
1-\texttt{alpha})

* Pg. 6 (and elsewhere): vectors with ldots should have commas before and
after the ldots.

* Pg. 7 second bullet "maxn.I=Plan" not clear what this means.

* Pg. 8 zi bullet: "must be between (inclusive of) the boundaries specified
for analysis i in x$lower$bound[i] and x$upper$bound[i]". Is this equivalent
to saying "must be in the interval x$lower$bound[i] \leq zi \leq x$upper$bound[i]"? If so, the latter construction is clearer.

* Pg. 8 second "theta" bullet: the two expressions "zi/sqrt(x$n.I[i]" and
"x$n.I[i]" should be formatted as either math or code--I'd suggest math for
the first and code for the second, but code for both is acceptable.

* Pg. 8, first sentence of section 4.4: "2-arm" s/b "two-arm"

* Pg. 9, "adj" bullet. "See details." Where are the details? If referring to
a help file, say something like, "See the xxx help file for details."

* Pg. 10,"pbar" bullet: \delta0 s/b \texttt{delta0}

* Pg. 10, second sentence of Section 5: add serial comma after print ("access, print, or plot")

* Pg. 11, n.I list item: following enumerated list should be nested (remove
\end{itemize} before and \begin{itemize} after)

* Pg. 12, 5 or "sf" bullet: "only" should modify "if", not "available" 
("available only if"

* Pg. 13, opening paragraph of Section 7: "potential interest" implies 
"to some", so those words can be stricken. "It is also possible for users"
can be shortened to "Users can", or, better, "You can", since presumably 
whoever is reading the manual is a user. Does "a function with a new spending 
function" say more than "a new spending function"? The "a" before "standard
spending functions" should be deleted.

* Pg. 13, opening paragraph of Section 7.1 (and elsewhere): equations (and 
inequalities) should be fully set in math mode: $\gamma = -4$, not $\gamma$ = -4

* Pg. 14, sentence beginning "Try this same plot using..." I found this hard to follow. Better to show the original code, then introduce the Kim-DeMets 
comparison and explicitly show the changed code. 

* Pg. 15, table caption: should be with table. Use table environment to 
surround tabular.

* Pg. 17, paragraph beginning "The logistic distribution": "logit" and "log"
should both be set in roman type, "log" by using "\log" and "logit" by using
"{\rm logit}" inside math mode.

* Pg. 18, paragraph beginning "This design now has slightly higher": "fourth 
interim was was saved" has extra "was"

* Pg. 20, first paragraph: "Note that sfupar and sflpar are specified in first"
add "the" before "first"




git-svn-id: svn://svn.r-forge.r-project.org/svnroot/gsdesign@118 edb9625f-4e0d-4859-8d74-9fd3b1da38cb